updating ui-servicepoints dependencies
Ui-servicepoints is currently using old versions of several dependencies,
including out-of-date versions of:
@folio/stripes,
@folio/react-intl-safe-html, and
react-intl

This is causing significant problems when you try to pull a local copy
of this module and use it in a development environment.  I've updated
the listed packages to more recent versions in the package-json, this
seems to have fixed the numerous errors this module was throwing when
used with the latest platform-core snapshot.