UISP-11: Build BigTest unit test code coverage above 80%
This update adds automated test coverage for this module.
there are two tests, one for each modal the handler displays.
the first test (accessModal) loads a dummy stripes-core instance
and generates a routing event to confirm that ui-sp will intercept
the event and generate a warning modal.  The second test just
mounts the servicepoint selection modal and checks to make sure
it displays servicepoint data correctly.  Actually changing the
users default service point is handled by stripes-core, and is
not tested here.

refs: https://issues.folio.org/browse/UISP-11