UISP-32 move login-response service-point parsing
Refactor the portions of the login-response parsing that deal with
optional interfaces (such as Service Points, which are part of the
`inventory` interface) out of stripes-core and into modules that do rely
on such interfaces.

This is just the first pass and only deals with the pieces that are
specifically linked to login and to reloading an existing session that
was save in local-storage. The logic that deals with the labeling of the
`ProfileDropdown` trigger and menu still knows about Service Points,
alas, will need additional cleanup.

This will need to be merged after the STCOR-649 work, which saves the
raw login response to session storage.

Refs UISP-32, STCOR-649