UISP-53 handle dynamically-defined servicepoints consumers
**BREAKING** When handling `SELECT_MODULE` events, inspect
`stripes.modules.servicepointsConsumer` to see if there is a value that
matches the destination application. `stripes.modules` is only available
beginning in `10.1.0`, hence why this is a breaking change.

This allows modules to add `servicepointsConsumer` to the
`stripes.actsAs` array in their `package.json` file and for this module
to discover that without any hard-coded coupling such as the legacy
hard-coded list of checkin, checkout, and requests.

Refs UISP-53